+++ docs/CHANGELOG.rst
@@ -25,6 +25,7 @@ Added
 - ``data_by_slug`` filter for jinja expression engine
 - Export RESOLWE_API_HOST environment variable in executor
 - Add ``check_installed()`` test utility function
+- Add support for configuring the network mode of Docker executor
 
 Changed
 -------